requirements.txt
requests==2.32.3
environs==14.1.1
python-telegram-bot==13.15
urllib3==1.26.15
logging==0.4.9.6